requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pgvector==0.2.4
celery==5.3.4
redis==5.0.1
langchain==0.2.16
langchain-openai==0.1.8
langchain-community==0.2.11
tenacity==8.2.3
crewai==0.36.0
openai>=1.26.0
ollama==0.1.7
unstructured==0.11.6
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2
python-magic==0.4.27
pillow==10.1.0
scikit-learn==1.3.2
numpy==1.26.4
python-dotenv==1.0.0
drf-spectacular==0.26.5
whitenoise==6.6.0
gunicorn==21.2.0
python-docx==1.1.0
pytesseract==0.3.10
lxml>=4.9.0
